pyproject.toml
[project]
name = "spec-agent-skills"
version = "0.1.0"
description = "Agent Skills specification as Strands Agents tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents>=1.27.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/spec_agent_skills"]
